ui: moved IStopViewer to package "model"
IStopViewer is referenced by IReviewConfigurable which is part of the
plug-in API. So IStopViewer must be moved to a package which is also
part of the plug-in API.
