+++ de.setsoftware.reviewtool.tests/src/de/setsoftware/reviewtool/config/ConfigurationInterpreterTest.java
@@ -27,10 +27,10 @@ import org.xml.sax.InputSource;
 
 import de.setsoftware.reviewtool.base.ReviewtoolException;
 import de.setsoftware.reviewtool.model.IReviewPersistence;
+import de.setsoftware.reviewtool.model.IStopViewer;
 import de.setsoftware.reviewtool.model.api.IChangeSource;
 import de.setsoftware.reviewtool.model.changestructure.IIrrelevanceDetermination;
 import de.setsoftware.reviewtool.preferredtransitions.api.IPreferredTransitionStrategy;
-import de.setsoftware.reviewtool.ui.IStopViewer;
 import de.setsoftware.reviewtool.ui.dialogs.EndReviewExtension;
 
 /**
